refactor(policy): GuardSpec is a data shape, move it into types
